+++ app/containers/AdsPage/constants.js
@@ -1,0 +1,3 @@
+export const LOAD_ADS = 'Ads/App/LOAD_ADS';
+export const LOAD_ADS_SUCCESS = 'boilerplate/App/LOAD_ADS_SUCCESS';
+export const LOAD_ADS_ERROR = 'boilerplate/App/LOAD_ADS_ERROR';
