+++ app/containers/AdsPage/saga.js
@@ -1,10 +1,9 @@
 import { put, takeLatest } from 'redux-saga/effects';
-import { LOAD_ADS } from 'containers/App/constants';
-import { adsLoaded, adsLoadingError } from 'containers/App/actions';
-
 import request from 'utils/request';
+import { LOAD_ADS } from './constants';
+import { adsLoaded, adsLoadingError } from './actions';
 
-export function* getAds() {
+export function* getAds(action) {
   try {
     const ads = yield request(`${GLOBAL_CONFIG.apiUrl}/random?nb_ads=50`);
     yield put(adsLoaded(ads));
