+++ app/containers/App/index.js
@@ -11,7 +11,7 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 
 import HomePage from 'containers/HomePage/Loadable';
-import Ads from 'containers/Ads';
+import AdsPage from 'containers/AdsPage';
 import NotFoundPage from 'containers/NotFoundPage/Loadable';
 import { Container, Footer } from 'react-bulma-components';
 import styled from 'styled-components';
@@ -43,7 +43,7 @@ export default function App() {
     <div>
       <Switch>
         <Route exact path="/" component={HomePage} />
-        <Route exact path="/ads" component={Ads} />
+        <Route path="/ads" component={AdsPage} />
         <Route component={NotFoundPage} />
       </Switch>
       <GlobalStyle />
