+++ app/index.html
@@ -12,7 +12,7 @@
   <meta name="mobile-web-app-capable" content="yes">
 
   <link rel="icon" href="/ads/favicon.ico" />
-  <title>React.js Boilerplate</title>
+  <title>Fight against disinformation</title>
 </head>
 
 <body>
